requirements.txt
opencv-python
numpy
pillow
flask
pytesseract
imwatermark
pymupdf
tensorflow
python-docx
sentence_transformers
piexif
sqlite3